requirements.txt
azure-core==1.30.1
azure-identity==1.16.0
azure-storage-blob==12.19.1
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.7
Flask==3.0.2
idna==3.7
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
msal==1.28.0
msal-extensions==1.1.0
msrest==0.7.1
oauthlib==3.2.2
packaging==24.0
portalocker==2.8.2
pycparser==2.22
PyJWT==2.8.0
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.1
